nav bar profile and recently viewed
#331
