+++ waltz-ng/client/navbar/directives/navbar-profile.html
@@ -1,0 +1,168 @@
+<ul class="nav navbar-nav navbar-right"
+    style="padding-right: 60px">
+
+    <li class="dropdown" uib-dropdown>
+        <a class="clickable"
+           uib-dropdown-toggle>
+                        <span waltz-if-anonymous="true">
+                            <waltz-icon name="user"></waltz-icon>
+                            Anonymous
+                            <b class="caret"></b>
+                        </span>
+            <span waltz-if-anonymous="false">
+                            <waltz-icon name="user"></waltz-icon>
+                            Profile
+                            <b class="caret"></b>
+                        </span>
+        </a>
+        <ul class="dropdown-menu"
+            uib-dropdown-menu
+            style="width: 250px; margin-right:8px"
+            role="menu"
+            class="no-overflow">
+
+            <li style="margin: 4px"
+                class="no-overflow">
+                            <span waltz-if-anonymous="false">
+                                <a ui-sref="main.profile.view ({userId: ctrl.user.userName})"
+                                   ng-bind="ctrl.user.userName"></a>
+                            </span>
+                <span waltz-if-anonymous="true">Anonymous</span>
+            </li>
+            <li style="margin: 4px"
+                ng-if="ctrl.allowDirectLogin"
+                class="no-overflow">
+                            <span waltz-if-anonymous="true">
+                                <a class="clickable"
+                                   ng-click="ctrl.login()">
+                                    <waltz-icon name="sign-in"></waltz-icon>
+                                    <span>Login</span>
+                                </a>
+                            </span>
+                <span waltz-if-anonymous="false">
+                                <a class="clickable"
+                                   ng-click="ctrl.logout()">
+                                <waltz-icon name="sign-out"></waltz-icon>
+                                Logout
+                            </a>
+                </span>
+
+            </li>
+
+            <li class="divider" waltz-has-role="ADMIN"></li>
+
+            <li style="margin: 4px"
+                ui-sref-active="side-active"
+                class="no-overflow">
+                            <span waltz-has-role="ADMIN">
+                                <a ui-sref="main.traits">
+                                    <waltz-icon class='text-muted small'
+                                                fixed-width="true"
+                                                name="cubes"></waltz-icon>
+                                    Manage Traits
+                                </a>
+                            </span>
+            </li>
+            <li style="margin: 4px"
+                ui-sref-active="side-active"
+                class="no-overflow">
+                            <span waltz-has-role="ADMIN">
+                                <a ui-sref="main.user.management">
+                                    <waltz-icon class='text-muted small'
+                                                fixed-width="true"
+                                                name="users"></waltz-icon>
+                                    Manage Users
+                                </a>
+                            </span>
+            </li>
+            <li style="margin: 4px"
+                ui-sref-active="side-active"
+                class="no-overflow">
+                            <span waltz-has-role="APP_EDITOR">
+                                <a ui-sref="main.app.registration">
+                                    <waltz-icon class='text-muted small'
+                                                fixed-width="true"
+                                                name="gift"></waltz-icon>
+                                    Register App
+                                </a>
+                            </span>
+            </li>
+            <li style="margin: 4px"
+                ui-sref-active="side-active"
+                class="no-overflow">
+                            <span waltz-has-role="ADMIN">
+                                <a ui-sref="main.settings">
+                                    <waltz-icon class='text-muted small'
+                                                fixed-width="true"
+                                                name="cog"></waltz-icon>
+                                    Settings
+                                </a>
+                            </span>
+            </li>
+            <li style="margin: 4px"
+                ui-sref-active="side-active"
+                class="no-overflow">
+                            <span waltz-has-role="ADMIN">
+                                <a ui-sref="main.source-data-ratings">
+                                    <waltz-icon class='text-muted small'
+                                                fixed-width="true"
+                                                name="map-signs"></waltz-icon>
+                                    Source Data Ratings
+                                </a>
+                            </span>
+            </li>
+            <li style="margin: 4px"
+                ui-sref-active="side-active"
+                class="no-overflow">
+                            <span waltz-has-role="ADMIN">
+                                <a ui-sref="main.user.log">
+                                    <waltz-icon class='text-muted small'
+                                                fixed-width="true"
+                                                name="list-ul"></waltz-icon>
+                                    User Logs
+                                </a>
+                            </span>
+            </li>
+
+        </ul>
+
+        <script type="text/ng-template" id="navbar/modal-login.html">
+            <div class="modal-header">
+                <h3 class="modal-title">
+                    Login to Waltz
+                    <small ng-bind="logoOverlayText"></small>
+                </h3>
+            </div>
+
+            <div class="modal-body">
+                <form>
+                    <label>Username</label>
+                    <input class='form-control input-sm'
+                           ng-model="username"
+                           autofocus
+                           placeholder="username">
+
+                    <label>Password</label>
+                    <input class='form-control input-sm'
+                           type="password"
+                           ng-model="password"
+                           placeholder="password">
+                </form>
+            </div>
+
+            <div class="modal-footer">
+                <button class="btn btn-primary"
+                        type="button"
+                        ng-click="ok()">
+                    OK
+                </button>
+                <button class="btn btn-warning"
+                        type="button"
+                        ng-click="cancel()">
+                    Cancel
+                </button>
+            </div>
+        </script>
+    </li>
+
+</ul>
