+++ waltz-ng/client/navbar/directives/navbar-recently-viewed.js
@@ -7,7 +7,7 @@ const template = require('./navbar-recently-viewed.html');
 
 
 const initialState = {
-    history: [],
+    history: []
 };
