+++ waltz-ng/client/navbar/index.js
@@ -15,6 +15,7 @@ export default (module) => {
     module
         .directive('waltzNavbar', require('./directives/navbar'))
         .directive('waltzNavbarRecentlyViewed', require('./directives/navbar-recently-viewed'))
+        .directive('waltzNavbarProfile', require('./directives/navbar-profile'))
         .directive('waltzNavSearchResults', require('./directives/nav-search-results'));
 
     module.component('waltzNavbarSearchForm', require('./components/navbar-search-form'));
